# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ApproveRequest:
    properties:
      comment:
        example: Document looks correct
        type: string
    type: object
...
  model.Document:
    properties:
      checked_by:
        type: string
      comment:
        type: string
      created_at:
        type: string
      file_url:
        type: string
      id:
        type: string
      sailor_id:
        type: string
      status:
        $ref: '#/definitions/model.DocumentStatus'
      type:
        type: string
      updated_at:
        type: string
    type: object
  model.DocumentStatus:
    enum:
    - pending
    - approved
    - rejected
    - need_revision
    type: string
    x-enum-varnames:
    - DocPending
    - DocApproved
    - DocRejected
    - DocNeedRevision
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Backend API for the sailing competition judging application. Handles
    regattas, race groups, individual race measurements, sailors, documents, and user
    authentication.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/user/judging-of-sailing
  title: Judging of Sailing API
  version: 1.0.0
paths:
...
  /documents:
    post:
      consumes:
      - application/json
      description: Uploads a new document (file reference) for a sailor. The file
        itself should be stored externally; this endpoint stores the reference URL.
      parameters:
      - description: Document details
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/model.Document'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Document'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload document
      tags:
      - documents
  /documents/{id}:
    delete:
      description: Deletes a document by ID.
      parameters:
      - description: Document UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete document
      tags:
      - documents
    get:
      description: Returns a single document by UUID.
      parameters:
      - description: Document UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Document'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: document not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get document by ID
      tags:
      - documents
  /documents/{id}/approve:
    post:
      description: Sets the document status to "approved". The checked_by field is
        set to the authenticated user.
      parameters:
      - description: Document UUID
        in: path
        name: id
        required: true
        type: string
      - description: Optional comment
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.ApproveRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve document
      tags:
      - documents
  /documents/{id}/reject:
    post:
      consumes:
      - application/json
      description: Sets the document status to "rejected" with a required comment
        explaining the reason.
      parameters:
      - description: Document UUID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason (comment is required)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ApproveRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject document
      tags:
      - documents
  /documents/{id}/request-revision:
    post:
      consumes:
      - application/json
      description: Sets the document status to "need_revision" with a comment explaining
        what needs to be changed.
      parameters:
      - description: Document UUID
        in: path
        name: id
        required: true
        type: string
      - description: Revision request (comment is required)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ApproveRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Request document revision
      tags:
      - documents
...
  /sailors/{sailor_id}/documents:
    get:
      description: Returns all documents for a given sailor, ordered by creation date
        descending.
      parameters:
      - description: Sailor UUID
        in: path
        name: sailor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Document'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List documents by sailor
      tags:
      - documents
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from POST /api/v1/auth/login. Use format: Bearer
      <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"